src/requirements.txt
Django==1.6.4
MySQL-python==1.2.3
PIL==1.1.7
Pillow==2.5.3
PyJWT==0.2.1
boto==2.19.0
django-crispy-forms==1.4.0
django-favicon==0.1.1
django-paypal==0.2.3
django-social-auth==0.7.28
django-storages==1.1.8
django-taggit==0.12.2
django-user-accounts==1.0b14
docutils==0.7
easy-thumbnails==2.1
egenix-mx-base==3.1.3
httplib2==0.9
netifaces==0.10.4
numpy==1.5.0
oauth2==1.5.211
oauthlib==0.6.3
pbr==1.10.0
psycopg2==2.5.4
pycrypto==2.3
python-openid==2.2.5
python-social-auth==0.2.1
rauth==0.7.2
reportlab==2.4
requests==2.6.2
requests-oauthlib==0.4.1
six==1.10.0
stevedore==1.17.1
stripe==1.30.0
virtualenv==13.0.1
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
wsgiref==0.1.2
